Rename configuration -> app-configuration
