+++ newswires/client/src/app-configuration.ts
